split sbml parser test
